refactor(GIL): move ilist trait implementation of InstBase from BasicBlock into InstBase file
