+++ lib/GIL/BasicBlock.cpp
@@ -87,21 +87,3 @@ void BasicBlock::setTerminator(InstBase *terminator)
 }
 
 } // end namespace glu::gil
-
-// ! TODO: implementation to move in InstBase file
-namespace llvm {
-glu::gil::BasicBlock *ilist_traits<glu::gil::InstBase>::getContainingBlock()
-{
-    size_t Offset = reinterpret_cast<size_t>(
-        &((glu::gil::BasicBlock *) nullptr
-              ->*glu::gil::BasicBlock::getSublistAccess(
-                  static_cast<glu::gil::InstBase *>(nullptr)
-              ))
-    );
-    iplist<glu::gil::InstBase> *Anchor
-        = static_cast<iplist<glu::gil::InstBase> *>(this);
-    return reinterpret_cast<glu::gil::BasicBlock *>(
-        reinterpret_cast<char *>(Anchor) - Offset
-    );
-}
-} // end namespace llvm
