+++ lib/GIL/InstBase.cpp
@@ -1,5 +1,5 @@
-#include "InstBase.hpp"
 #include "InstVisitor.hpp"
+#include "BasicBlock.hpp"
 
 #include <llvm/ADT/StringRef.h>
 
@@ -24,4 +24,21 @@ BasicBlock *Value::getDefiningBlock() const
     return value.get<InstBase *>()->getParent();
 }
 
-} // namespace glu::gil
+} // end namespace glu::gil
+
+namespace llvm {
+glu::gil::BasicBlock *ilist_traits<glu::gil::InstBase>::getContainingBlock()
+{
+    size_t Offset = reinterpret_cast<size_t>(
+        &((glu::gil::BasicBlock *) nullptr
+              ->*glu::gil::BasicBlock::getSublistAccess(
+                  static_cast<glu::gil::InstBase *>(nullptr)
+              ))
+    );
+    iplist<glu::gil::InstBase> *Anchor
+        = static_cast<iplist<glu::gil::InstBase> *>(this);
+    return reinterpret_cast<glu::gil::BasicBlock *>(
+        reinterpret_cast<char *>(Anchor) - Offset
+    );
+}
+} // end namespace llvm
